Refactor and extend simulation and option chain logic
Reorganized imports, adjusted formatting, and introduced new methods for `WalkParams` and `OptionData`. Added extensive comments and validation mechanisms for `OptionData`. Integrated new test cases and enhanced random walk functionality.```
