+++ src/strategies/iron_condor.rs
@@ -1869,7 +1869,7 @@ mod tests_iron_condor_strategies {
 #[cfg(test)]
 mod tests_iron_condor_optimizable {
     use super::*;
-    use crate::chains::chain::OptionData;
+    use crate::chains::OptionData;
     use crate::model::types::ExpirationDate;
     use crate::pos;
     use crate::spos;
