+++ tests/unit/chain/mod.rs
@@ -1,0 +1,7 @@
+/******************************************************************************
+    Author: Joaquín Béjar García
+    Email: jb@taunais.com 
+    Date: 27/3/25
+ ******************************************************************************/
+ 
+mod random_walk_chain;
