+++ tests/unit/chain/random_walk_chain.rs
@@ -1,24 +1,27 @@
-use optionstratlib::chains::chain::OptionChain;
+/******************************************************************************
+   Author: Joaquín Béjar García
+   Email: jb@taunais.com
+   Date: 27/3/25
+******************************************************************************/
+use optionstratlib::chains::OptionChain;
 use optionstratlib::simulation::randomwalk::RandomWalk;
 use optionstratlib::simulation::steps::{Step, Xstep, Ystep};
 use optionstratlib::simulation::{WalkParams, WalkType, WalkTypeAble};
-use optionstratlib::utils::setup_logger;
-use optionstratlib::utils::time::{TimeFrame, convert_time_frame, get_x_days_formatted};
-use optionstratlib::visualization::utils::{Graph, GraphBackend};
+use optionstratlib::utils::time::{convert_time_frame, get_x_days_formatted};
+use optionstratlib::utils::{Len, TimeFrame, setup_logger};
 use optionstratlib::{ExpirationDate, Positive, pos};
 use rust_decimal_macros::dec;
-use tracing::{debug, info};
+use std::error::Error;
+use tracing::info;
 
-#[warn(dead_code)]
-struct Walker {}
-
-impl Walker {
+struct MockWalker {}
+impl MockWalker {
     fn new() -> Self {
-        Walker {}
+        MockWalker {}
     }
 }
 
-impl WalkTypeAble<Positive, OptionChain> for Walker {}
+impl WalkTypeAble<Positive, OptionChain> for MockWalker {}
 
 fn create_chain_from_step(
     previous_y_step: &Ystep<OptionChain>,
@@ -33,7 +36,7 @@ fn create_chain_from_step(
     }
 
     let new_chain = OptionChain::build_chain(&chain_params);
-
+    assert!(new_chain.len() > 0,);
     Ok(new_chain)
 }
 
@@ -42,7 +45,6 @@ fn generator(walk_params: &WalkParams<Positive, OptionChain>) -> Vec<Step<Positi
     let mut y_steps = walk_params.walker.geometric_brownian(&walk_params).unwrap();
     let _ = y_steps.remove(0);
     let mut steps: Vec<Step<Positive, OptionChain>> = vec![walk_params.init_step.clone()];
-
     let mut previous_x_step = walk_params.init_step.x.clone();
     let mut previous_y_step = walk_params.ystep();
 
@@ -59,24 +61,31 @@ fn generator(walk_params: &WalkParams<Positive, OptionChain>) -> Vec<Step<Positi
             x: previous_x_step,
             y: previous_y_step.clone(),
         };
+        println!("{}", step);
+        assert!(*step.x.index() >= 1 );
+        assert_eq!(*step.x.time_unit(), TimeFrame::Minute);
+        assert_eq!(*step.x.step_size_in_time(), Positive::ONE);
+        assert!(*step.y.index() >= 1 );
         steps.push(step)
     }
-
+    
     assert!(steps.len() <= walk_params.size);
     steps
 }
-
-fn main() -> Result<(), Box<dyn std::error::Error>> {
+#[test]
+fn test_random_walk_chain() -> Result<(), Box<dyn Error>> {
     setup_logger();
-    let n_steps = 43_200; // 30 days in minutes
-    // let n_steps = 10;
+    let n_steps = 20;
+
     let mut initial_chain =
         OptionChain::load_from_json("examples/Chains/SP500-18-oct-2024-5781.88.json")?;
     initial_chain.update_expiration_date(get_x_days_formatted(2));
-    // info!("Initial Chain: {}", initial_chain);
+
+    assert_eq!(initial_chain.underlying_price, pos!(5781.88));
+    assert_eq!(initial_chain.symbol, "SP500");
 
     let std_dev = pos!(20.0);
-    let walker = Box::new(Walker::new());
+    let walker = Box::new(MockWalker::new());
     let days = pos!(30.0);
 
     let walk_params = WalkParams {
@@ -94,15 +103,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     };
 
     let random_walk = RandomWalk::new("Random Walk".to_string(), &walk_params, generator);
-    debug!("Random Walk: {}", random_walk);
+    info!("Random Walk: {}", random_walk);
+    assert_eq!(random_walk.len(), n_steps);
 
-    random_walk.graph(
-        GraphBackend::Bitmap {
-            file_path: "Draws/Simulation/random_walk_chain.png",
-            size: (1200, 800),
-        },
-        20,
-    )?;
     info!("Last Chain: {}", random_walk.last().unwrap().y.value());
 
     Ok(())
