+++ tests/unit/mod.rs
@@ -5,3 +5,5 @@
 ******************************************************************************/
 
 mod strategies;
+
+mod chain;
