+++ tests/unit/strategies/delta/strategy_bear_call_spread.rs
@@ -10,7 +10,6 @@ use rust_decimal_macros::dec;
 use std::error::Error;
 
 #[test]
-
 fn test_bear_call_spread_integration() -> Result<(), Box<dyn Error>> {
     setup_logger();
     // Define inputs for the BearCallSpread strategy
